Add files via upload
WebScraper GUI is a graphical user interface application that enables users to easily scrape data from any website. With a simple and intuitive interface, users can input the URL of the website they want to scrape, and the program will automatically extract all the content from the website and display it in a scrollable text box. The application also features buttons to reset, clear, enter, and exit the program, making it easy to use for both novice and advanced users. It can be used for various purposes, such as data collection, web scraping for personal use, or even educational purposes. It's a time saver and a practical tool for web scraping tasks